separated iex audioexport module
