+++ src/importexport/audioexport/CMakeLists.txt
@@ -17,25 +17,19 @@
 #  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #=============================================================================
 
-# We are separating the importexport module into separate modules,
-# soon there will be nothing left of it
-set(MODULE importexport)
+set(MODULE iex_audioexport)
 
 set(MODULE_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/importexportmodule.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importexportmodule.h
-    ${CMAKE_CURRENT_LIST_DIR}/iimportexportconfiguration.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/importexportconfiguration.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/importexportconfiguration.h
-    
-    ${CMAKE_CURRENT_LIST_DIR}/internal/svgwriter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/svgwriter.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/svggenerator.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/svggenerator.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/pngwriter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/pngwriter.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/pdfwriter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/pdfwriter.h
+    ${CMAKE_CURRENT_LIST_DIR}/audioexportmodule.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/audioexportmodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/mp3writer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/mp3writer.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/wavewriter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/wavewriter.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/oggwriter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/oggwriter.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/flacwriter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/flacwriter.h
     )
 
 set(MODULE_LINK
